Improve NaN handling in Optimizer
(1) print out if optimizer iteration stopped on NaN (2) handle NaN's in cost function during line search. These should usually not happen, but for e.g. the SSNP forward model they can show up if the initial step is much too large. Currently handling these in the base class, but if this proves not to be useful in general this could be shifted to field_prop.py SSNP